requirements.linux.txt
matplotlib
torch==2.4.0
pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-2.4.0+cu124.html
torch_geometric
torch_geometric_temporal
google-cloud-storage
google-cloud-logging
google-cloud-secret-manager
cloudml-hypertune
wandb
scikit-learn
numpy==1.26.4
